Removed double error throw for subscribeFail
Currently, if a sub fails, it calls then subscribeFail callback, which is good. However, It additionally emits an `error` as well, which is bad, since if I wanted to know about that error, I would subscribe to the `subscribeFail`. Additionally, the error it emits is just a string (bad), whereas all other errors are error objects (good). 

Pretty sure this is a bug, so I closed it up.