Add isAuthenticated to state
Whenever a client calls the authenticate method, or receives a
successful handshake, or gets a #setAuthToken, the auth state is set to
true.
When a client is disconnected or an auth token is removed, or an auth
token expires, the isAuthorized state turns false.

Also, performs check for malformed tokens before sending them to the
server.