Make client disconnect after number of attempts
Just thought it would be great to have an option on client side that defines the maximum number of attempts for the client to keep reconnecting. ( I set 1000 attempts as default. ) . Is it something that could be implemented in the sc-client or is it better approach to have this handled inside my app instead?