v7.0.0 - socket.authState can no longer be in a 'pending' state.
The 'pending' authState used to represent the ambiguous authentication state which might occur when the
socket is disconnected (for whatever reason). It was removed because it was too ambiguous to be useful.
Instead, socket.authState now reports the last known authentication status (optimistic approach).
Once a socket is authenticated with a valid JWT, the socket will now report itself as being 'authenticated'
until the server explicitly indicates otherwise (e.g. the server invalidated the token).